distinguish between ui and unit tests (unittests not ready)
