+++ .gitignore
@@ -11,7 +11,7 @@
 # production
 /build
 /dist
-/test/build
+/test/**/build
 
 # misc
 .DS_Store
