+++ jest.config.js
@@ -1,8 +1,12 @@
 module.exports = {
-  collectCoverageFrom: ['src/**/*.{js,jsx,ts,tsx}', '!src/**/*.d.ts'],
+  collectCoverageFrom: ['test/**/*.{js,jsx,ts,tsx}', '!src/**/*.d.ts'],
   setupFiles: ['react-app-polyfill/jsdom'],
   setupFilesAfterEnv: [],
-  testMatch: ['<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}', '<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}'],
+  testMatch: [
+    '<rootDir>/test/unittests/**/__tests__/**/*.{js,jsx,ts,tsx}',
+    '<rootDir>/test/unittests/**/*.{spec,test}.{js,jsx,ts,tsx}',
+    '<rootDir>/test/unittests/**/*.{spec,test}.{js,jsx,ts,tsx}',
+  ],
   testEnvironment: 'jest-environment-jsdom-fourteen',
   transform: {
     '^.+\\.(js|jsx|ts|tsx)$': '<rootDir>/node_modules/babel-jest',
