+++ package.json
@@ -66,6 +66,7 @@
     "style-loader": "0.23.1",
     "styled-components": "^4.4.0",
     "terser-webpack-plugin": "1.2.3",
+    "ts-jest": "^24.1.0",
     "ts-loader": "^6.0.4",
     "ts-pnp": "1.1.2",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
@@ -102,7 +103,7 @@
   },
   "scripts": {
     "dev": "concurrently \"yarn frontend-dev\" \"yarn backend-dev\"",
-    "test": "npm run test-build && npm run test-backend",
+    "test": "npm run test-build && npm run test-ui",
     "production": "npm run frontend-build && npm run backend-build && electron-builder .",
     "frontend-dev": "node scripts/frontend/start.js",
     "frontend-build": "node scripts/frontend/build.js",
@@ -110,7 +111,8 @@
     "backend-build": "node scripts/backend/build.js",
     "backend-dev": "node scripts/backend/start.js",
     "test-build": "node scripts/backend/test.js && node scripts/frontend/build.js && npm run lint &&tsc --build tsconfig.test.json",
-    "test-backend": "mocha",
+    "test-ui": "mocha",
+    "test-unit": "jest ",
     "lint": "tslint 'backend/**/*.{ts,tsx}' 'src/**/*.{ts,tsx}' 'test/**/*.{ts,tsx}'"
   },
   "eslintConfig": {
