+++ test/mocha.opts
@@ -1,3 +1,3 @@
-test/build/test/**/*.js
+test/uitests/build/**/*.js
 --exit
 --require @babel/register
